requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.7
django-environ==0.4.5
djangorestframework==3.12.2
environ==1.0
fsspec==2021.4.0
gunicorn==20.0.4
idna==2.10
install==1.3.4
mypy-extensions==0.4.3
numpy==1.20.2
pandas==1.2.4
pathspec==0.8.1
phonenumbers==8.12.21
pyarrow==3.0.0
pydantic==1.8.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
regex==2021.3.17
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
